requirements.txt
weaviate-client==3.24.1
PyPDF2==3.0.1
sentence-transformers==2.2.2
tqdm==4.66.1
python-dotenv==1.0.0